% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{BLUERED}
\alias{BLUERED}
\title{A dataset used by}
\format{
## `BLUERED`
}
\source{
unknown
}
\usage{
BLUERED
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
