% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{COLLECTIONS}
\alias{COLLECTIONS}
\title{A dataset used by}
\format{
## `COLLECTIONS`
}
\source{
unknown
}
\usage{
COLLECTIONS
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
