% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{COLORS}
\alias{COLORS}
\title{A dataset used by}
\format{
## `COLORS`
}
\source{
unknown
}
\usage{
COLORS
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
