% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{FAMILIES}
\alias{FAMILIES}
\title{A dataset used by}
\format{
## `FAMILIES`
}
\source{
unknown
}
\usage{
FAMILIES
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
