% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{GENE_SUMMARY}
\alias{GENE_SUMMARY}
\title{A dataset used by}
\format{
## `GENE_SUMMARY`
}
\source{
unknown
}
\usage{
GENE_SUMMARY
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
