% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{GENE_SYMBOL}
\alias{GENE_SYMBOL}
\title{A dataset used by}
\format{
## `GENE_SYMBOL`
}
\source{
unknown
}
\usage{
GENE_SYMBOL
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
