% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_data}
\alias{get_data}
\title{Get a playdata dataset dynamically from a variable name}
\usage{
get_data(name)
}
\arguments{
\item{name}{string. Name of dataset.}
}
\description{
This function allows users to access playdata datasets when the dataset to
be accessed is only known dynamically or through a variable name
}
\examples{
all.equal(get_dataset('BLUERED'), playdata::BLUERED)
}
