% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{getGSETS}
\alias{getGSETS}
\title{A dataset used by}
\format{
## `getGSETS`
}
\source{
unknown
}
\usage{
getGSETS
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
