% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{GRAPHITE_PPI}
\alias{GRAPHITE_PPI}
\title{PPI dataframe containing from-to edges from mixed
(metabolite+protein) pathways downloaded fromGraphite.}
\format{
## `data.frame`
}
\source{
https://www.ebi.ac.uk/chebi/
}
\usage{
GRAPHITE_PPI
}
\description{
PPI dataframe containing from-to edges from mixed
(metabolite+protein) pathways downloaded fromGraphite.
}
\keyword{datasets}
