% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{GSET_GENES}
\alias{GSET_GENES}
\title{A dataset used by}
\format{
## `GSET_GENES`
}
\source{
unknown
}
\usage{
GSET_GENES
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
