% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{GSET_INFO}
\alias{GSET_INFO}
\title{Information for database source of Omics Playground Genesets}
\format{
## list containign data.frames with metadata for databases
}
\source{
MSigDB (https://www.gsea-msigdb.org/gsea/msigdb) and  enrichR (https://maayanlab.cloud/Enrichr/)
}
\usage{
GSET_INFO
}
\description{
Information for database source of Omics Playground Genesets
}
\keyword{datasets}
