% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{GSET_PREFIX_REGEX}
\alias{GSET_PREFIX_REGEX}
\title{A dataset used by}
\format{
## `GSET_PREFIX_REGEX`
}
\source{
unknown
}
\usage{
GSET_PREFIX_REGEX
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
