% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{GSET_SIZE}
\alias{GSET_SIZE}
\title{Raw number of genes per geneset}
\format{
## `GSET_SIZE`
A numeric vector of number of genes per geneset, names attribute refer to the geneset name.
}
\source{
unknown
}
\usage{
GSET_SIZE
}
\description{
Raw number of genes per geneset
}
\keyword{datasets}
