% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{GSETxGENE}
\alias{GSETxGENE}
\title{A collection of genesets used in Omics Playground}
\format{
## `GSET_SPARSEG_XL`
A dgCMatrix / Matrix object
}
\source{
unknown
}
\usage{
GSETxGENE
}
\description{
Geneset in rows, genes as columns.
}
\keyword{datasets}
