% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{iGSETS}
\alias{iGSETS}
\title{A dataset used by}
\format{
## `iGSETS`
}
\source{
unknown
}
\usage{
iGSETS
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
