% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{IMMPROT}
\alias{IMMPROT}
\title{A dataset used by}
\format{
## `IMMPROT`
}
\source{
unknown
}
\usage{
IMMPROT
}
\description{
No idea what this data is but it is used
}
\keyword{datasets}
