% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{IMMPROT_SIGNATURE1000}
\alias{IMMPROT_SIGNATURE1000}
\title{A dataset used by}
\format{
## `IMMPROT_SIGNATURE1000`
}
\source{
unknown
}
\usage{
IMMPROT_SIGNATURE1000
}
\description{
No idea what this data is but it is used
}
\keyword{datasets}
