% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{L1000_ACTIVITYS_N20D1011}
\alias{L1000_ACTIVITYS_N20D1011}
\title{A dataset used}
\format{
## `L1000_ACTIVITYS_N20D1011`
}
\source{
unknown
}
\usage{
L1000_ACTIVITYS_N20D1011
}
\description{
No idea what this data is
}
\keyword{datasets}
