% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{METABOLITE_ANNOTATION}
\alias{METABOLITE_ANNOTATION}
\title{Annotation table of metabolites}
\format{
## `data.frame`
}
\source{
https://www.metaboanalyst.ca/ and https://www.ebi.ac.uk/chebi/
}
\usage{
METABOLITE_ANNOTATION
}
\description{
Annotation table of metabolites
}
\keyword{datasets}
