% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{METABOLITE_ID}
\alias{METABOLITE_ID}
\title{ID conversion table of metabolites}
\format{
## `data.frame`
}
\source{
https://www.metaboanalyst.ca/ and https://www.ebi.ac.uk/chebi/
}
\usage{
METABOLITE_ID
}
\description{
ID conversion table of metabolites
}
\keyword{datasets}
