% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{METABOLITE_METADATA}
\alias{METABOLITE_METADATA}
\title{Metabolite metadata information}
\format{
## `data.frame`
}
\source{
https://www.ebi.ac.uk/chebi/
}
\usage{
METABOLITE_METADATA
}
\description{
Metabolite metadata information
}
\keyword{datasets}
