% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{MSETxMETABOLITE}
\alias{MSETxMETABOLITE}
\title{Sparse matrix with rows as metabolic pathways and columns as metabolites.}
\format{
## `gmt`
}
\source{
Nishida K (2021). AHWikipathwaysDbs: Metabolites linked to WikiPathways pathways (for AnnotationHub). R package version 0.99.4, https://github.com/kozo2/AHWikipathwaysDbs and https://www.ebi.ac.uk/chebi/
}
\usage{
MSETxMETABOLITE
}
\description{
Combined gmt file of reactome metabolites (metabolic pathways vs. chebi ids) and gmt file of WikiPathways metabolites (metabolic pathways vs. chebi ids) as sparse matrix (rows as metabolic pathways and columns as metabolites).
}
\keyword{datasets}
