% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{PHOSPHOSITE}
\alias{PHOSPHOSITE}
\title{A dataset used by}
\format{
## `PHOSPHOSITE`
}
\source{
unknown
}
\usage{
PHOSPHOSITE
}
\description{
Phosphosite database downloaded from dbPaf and dbPsp
}
\keyword{datasets}
