% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{PURPLEYELLOW}
\alias{PURPLEYELLOW}
\title{A dataset used by}
\format{
## `PURPLEYELLOW`
}
\source{
unknown
}
\usage{
PURPLEYELLOW
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
