% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{REACTOME_METABOLITES}
\alias{REACTOME_METABOLITES}
\title{gmt file of reactome metabolites (metabolic pathways vs. chebi ids)}
\format{
## `gmt`
}
\source{
https://reactome.org/
}
\usage{
REACTOME_METABOLITES
}
\description{
gmt file of reactome metabolites (metabolic pathways vs. chebi ids)
}
\keyword{datasets}
