% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{RTCGA_SURVIVAL}
\alias{RTCGA_SURVIVAL}
\title{A dataset used by}
\format{
## `RTCGA_SURVIVAL`
}
\source{
unknown
}
\usage{
RTCGA_SURVIVAL
}
\description{
No idea what this data is but it is used .
}
\keyword{datasets}
