% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{SIGNATURE_IMMUNEMETA}
\alias{SIGNATURE_IMMUNEMETA}
\title{A dataset used by `compute_deconvolution`}
\format{
## `SIGNATURE_IMMUNEMETA`
}
\source{
unknown
}
\usage{
SIGNATURE_IMMUNEMETA
}
\description{
No idea what this data is but it is used in `compute_deconvolution`.
}
\keyword{datasets}
