% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{TISSUE}
\alias{TISSUE}
\title{A dataset used by}
\format{
## `TISSUE`
}
\source{
unknown
}
\usage{
TISSUE
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
