% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{TISSUE_GRP}
\alias{TISSUE_GRP}
\title{A dataset used by}
\format{
## `TISSUE_GRP`
}
\source{
unknown
}
\usage{
TISSUE_GRP
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
