% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{WP_METABOLITES}
\alias{WP_METABOLITES}
\title{gmt file of WikiPathways metabolites (metabolic pathways vs. chebi ids)}
\format{
## `gmt`
}
\source{
Nishida K (2021). AHWikipathwaysDbs: Metabolites linked to WikiPathways pathways (for AnnotationHub). R package version 0.99.4, https://github.com/kozo2/AHWikipathwaysDbs.
}
\usage{
WP_METABOLITES
}
\description{
gmt file of WikiPathways metabolites (metabolic pathways vs. chebi ids)
}
\keyword{datasets}
