% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdata-data.R
\docType{data}
\name{GSETxGENE}
\alias{GSETxGENE}
\title{A dataset used by}
\format{
## `GSETxGENE`
}
\source{
unknown
}
\usage{
GSETxGENE
}
\description{
No idea what this data is but it is used.
}
\keyword{datasets}
